Refactor import statements in controllers to streamline utility access
- Updated multiple controller files to import the extractUserId function from a centralized utils index instead of individual files, enhancing code organization and maintainability.
